+++ packages/calcite-components/src/components/tab-title/resources.ts
@@ -16,4 +16,4 @@ export const CSS = {
 
 export const ICONS = {
   close: "x",
-};
+} as const;