+++ packages/calcite-components-angular/projects/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.10.2-next.2...@esri/calcite-components-angular@2.11.0-next.0) (2024-07-02)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
+## [2.10.2-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.10.2-next.1...@esri/calcite-components-angular@2.10.2-next.2) (2024-07-01)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
 ## [2.10.2-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.10.2-next.0...@esri/calcite-components-angular@2.10.2-next.1) (2024-07-01)
 
 **Note:** Version bump only for package @esri/calcite-components-angular